pyproject.toml
[tool.poetry]
name = "euro-cert-task"
version = "0.1.0"
description = ""
authors = ["lukaszklekowski <lukasz.k.klekowski@gmail.com>"]
readme = "README.md"
package-mode=false

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^7.2.1"
Django = "5.0.6"
djangorestframework-camel-case = "1.4.2"
djangorestframework = "3.15.1"
dj-rest-auth = "5.0.2"


[tool.poetry.group.dev.dependencies]
pytest-watch = "^4.2.0"
pdbpp = "^0.10.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
